feat: add http pauses prevent hitting API too hard

+++ sdk/src/main/java/io/customer/sdk/CustomerIOConfig.kt
@@ -20,5 +20,5 @@ class CustomerIOConfig(
      * The number of seconds to delay running queue after a task has been added to it.
      * We do not recommend modifying this value because it impacts battery life of mobile device.
      */
-    val backgroundQueueSecondsDelay: Long
+    val backgroundQueueSecondsDelay: Double
 )
